PlayList_Builder/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3

# Data / ML
scikit-learn==1.5.2
joblib==1.4.2
numpy==2.1.1
scipy==1.14.1
pandas==2.2.3

# UI
streamlit==1.38.0

# Dev tools
black==24.8.0  #python code formatter
isort==5.13.2   #A tool to sort and group your import statements.
